requirements.txt
python-telegram-bot[job-queue]
google-generativeai
python-dotenv
pypdf
python-docx
pytesseract
pillow
pdfplumber
fastapi
uvicorn
requests
regex
